// src/app/client/model/cart.ts
/**
 * my api
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { CartItem } from './cartItem';


export class Cart {
    id?: string;
    items?: Array<CartItem> = new Array<CartItem>();
    deliveryPrice?: number;
    itemsTotal?: number;
    timestamp?: number;
}