// src/app/client/model/cartItem.ts
/**
 * my api
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { Product } from './product';


export class CartItem {
    productId?: string;
    variationId?: string;
    product?: Product;
    available?: number;
    withoutCount?: boolean;
    title?: string;
    count?: number;
    price?: number;
}